+++ CHANGELOG.md
@@ -8,6 +8,12 @@
 
 ## v0.0
 
+### v0.0.0-18
+- 🟥 rework `Thumbprint` AGAIN (i'm going insane)
+  - changed names of options that nobody will use
+  - `preview` is the bytename part of default 8 bytes
+  - `sigil` is the bytename part of default 4 bytes
+
 ### v0.0.0-17
 - 🟥 adjust `Thumbprint` to use 6-byte sigils by default, hah